Eliminate build warnings, clean packaging, and update dependencies
- Fix all -Xlint warnings (serial, deprecation, redundant/lossy casts,
  try-with-resources, InterruptedException-throwing close()).
- Make AcpCommandTest deterministic regardless of host sandbox backend.
- Make `mvn clean package` warning-free: scope the benchmarks shade to the
  ajent-cli uber jar + JMH runtime, drop module-info/manifest overlaps, merge
  license/notice, skip the empty ajent-parity jar, disable annotation
  processing for benchmarks tests, and quiet CycloneDX schema-keyword logs.
- Bump JUnit 6.1.0 to 6.1.2, maven-compiler-plugin 3.14.1 to 3.15.0, and
  JLine 3.30.0 to 4.3.1 (migrate to Size.of / getSize() API).

+++ ajent-core/src/main/java/com/github/skanga/ajent/core/workspace/WorkspaceMatcher.java
@@ -88,7 +88,7 @@ public final class WorkspaceMatcher {
     }
     if (needleIndex < needle.length()) return Integer.MIN_VALUE;
     score -= skipped;
-    score -= 2 * path.chars().filter(value -> value == '/' || value == '\\').count();
+    score -= 2 * (int) path.chars().filter(value -> value == '/' || value == '\\').count();
     return score + classBias(path) + bonus;
   }
